rename ChartExamples.py to match convention
