+++ controller/DES/figure_list_select.py
@@ -4,7 +4,7 @@ figure list controller
 import sys
 sys.dont_write_bytecode = True
 import PySimpleGUI as sg
-import view.ChartExamples as ce
+import view.chart_examples as ce
 
 def accept(event, values, state):
     view = state['view']
