+++ view/chart_examples.py
